refactor: moves bin to src
