+++ .github/codeql/codeql-config.yml
@@ -1,7 +1,6 @@
 name: 'CodeQL config'
 
 paths:
-  - bin/
   - dist/esm/
   - src/
   - tools/
