--- bin/cli.js
@@ -1,5 +1,0 @@
-#!/usr/bin/env node
-
-import { cli } from "../dist/esm/cli.js";
-
-await cli();
