+++ package.json
@@ -19,7 +19,7 @@
     "url": "https://sverweij.github.io"
   },
   "license": "MIT",
-  "bin": "bin/cli.js",
+  "bin": "dist/esm/cli.js",
   "main": "dist/esm/main.js",
   "module": "dist/esm/main.js",
   "type": "module",
@@ -34,7 +34,6 @@
   },
   "types": "types/watskeburt.d.ts",
   "files": [
-    "bin",
     "dist",
     "!**/*.DS_Store",
     "types",
@@ -76,22 +75,22 @@
     "test": "node --no-warnings --loader 'ts-node/esm' --test-reporter dot --test src/*.spec.ts src/**/*.spec.ts",
     "test:cover": "c8 npm test",
     "test:only-for-node-16-without-the-test-reporter": "node --no-warnings --loader 'ts-node/esm' --test src/*.spec.ts src/**/*.spec.ts",
-    "depcruise": "depcruise bin dist src types",
-    "depcruise:graph": "depcruise bin src types --include-only '^(bin|dist|src|types)' --output-type dot | dot -T svg | tee docs/dependency-graph.svg | depcruise-wrap-stream-in-html > docs/dependency-graph.html",
-    "depcruise:graph:archi": "depcruise bin src --include-only '^(bin|dist|src|types)' --output-type archi | dot -T svg | depcruise-wrap-stream-in-html > docs/high-level-dependency-graph.html",
-    "depcruise:graph:dev": "depcruise bin dist src types --include-only '^(bin|dist|src|types)' --prefix vscode://file/$(pwd)/ --output-type dot | dot -T svg | depcruise-wrap-stream-in-html | browser",
-    "depcruise:graph:diff:dev": "depcruise bin dist src types --include-only '^(bin|dist|src|types)' --highlight \"$(node bin/cli.js main -T regex)\" --prefix vscode://file/$(pwd)/ --output-type dot | dot -T svg | depcruise-wrap-stream-in-html | browser",
-    "depcruise:graph:diff:mermaid": "depcruise bin dist src types --include-only '^(bin|dist|src|types)' --output-type mermaid --output-to - --highlight \"$(node bin/cli.js $SHA -T regex)\"",
-    "depcruise:html": "depcruise bin src types --progress --output-type err-html --output-to dependency-violation-report.html",
-    "depcruise:text": "depcruise bin src types --progress --output-type text",
-    "depcruise:focus": "depcruise bin src types --progress --output-type text --focus",
-    "depcruise:reaches": "depcruise bin src types --progress --output-type text --reaches",
-    "format": "prettier --write \"{bin,src,tools}/**/*.{js,ts}\" \"types/**/*.ts\" \"*.{json,yml,md,js}\"",
-    "format:check": "prettier --log-level warn --check \"{bin,src,tools}/**/*.ts\" \"types/**/*.ts\" \"*.{json,yml,md,js}\"",
+    "depcruise": "depcruise dist src types",
+    "depcruise:graph": "depcruise src types --include-only '^(dist|src|types)' --output-type dot | dot -T svg | tee docs/dependency-graph.svg | depcruise-wrap-stream-in-html > docs/dependency-graph.html",
+    "depcruise:graph:archi": "depcruise src --include-only '^(dist|src|types)' --output-type archi | dot -T svg | depcruise-wrap-stream-in-html > docs/high-level-dependency-graph.html",
+    "depcruise:graph:dev": "depcruise dist src types --include-only '^(dist|src|types)' --prefix vscode://file/$(pwd)/ --output-type dot | dot -T svg | depcruise-wrap-stream-in-html | browser",
+    "depcruise:graph:diff:dev": "depcruise dist src types --include-only '^(dist|src|types)' --highlight \"$(node dist/esm/cli.js main -T regex)\" --prefix vscode://file/$(pwd)/ --output-type dot | dot -T svg | depcruise-wrap-stream-in-html | browser",
+    "depcruise:graph:diff:mermaid": "depcruise dist src types --include-only '^(dist|src|types)' --output-type mermaid --output-to - --highlight \"$(node dist/esm/cli.js $SHA -T regex)\"",
+    "depcruise:html": "depcruise src types --progress --output-type err-html --output-to dependency-violation-report.html",
+    "depcruise:text": "depcruise src types --progress --output-type text",
+    "depcruise:focus": "depcruise src types --progress --output-type text --focus",
+    "depcruise:reaches": "depcruise src types --progress --output-type text --reaches",
+    "format": "prettier --write \"{src,tools}/**/*.{js,ts}\" \"types/**/*.ts\" \"*.{json,yml,md,js}\"",
+    "format:check": "prettier --log-level warn --check \"{src,tools}/**/*.ts\" \"types/**/*.ts\" \"*.{json,yml,md,js}\"",
     "lint": "npm-run-all --parallel --aggregate-output format:check lint:eslint lint:types",
     "lint:fix": "npm-run-all --parallel --aggregate-output format lint:eslint:fix",
-    "lint:eslint": "eslint bin src types tools --cache --cache-location node_modules/.cache/eslint/",
-    "lint:eslint:fix": "eslint bin src types tools --fix --cache --cache-location node_modules/.cache/eslint/",
+    "lint:eslint": "eslint src types tools --cache --cache-location node_modules/.cache/eslint/",
+    "lint:eslint:fix": "eslint src types tools --fix --cache --cache-location node_modules/.cache/eslint/",
     "lint:types": "tsc --noEmit",
     "scm:stage": "git add .",
     "update-dependencies": "run-s upem:update upem:install lint:fix check",
